\name{fArma-package}


\alias{fArma-package}
\alias{fArma}


\docType{package}


\title{Modelling ARMA Time Series Processes}


\description{

    The Rmetrics "fArma" package is a collection of functions to 
    analyze and model ARMA time series processes which special emphesis
    in Finance.
    
}


\details{

    \preformatted{
        Package:    \tab fArma\cr
        Type:       \tab Package\cr
        Version:    \tab R 3.0.1\cr
        Date:       \tab 2014\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2014 Rmetrics Assiciation\cr
        URL:        \tab \url{https://www.rmetrics.org}
    }
    
}


\keyword{models}


